Create Account.java
A Bank
Look ¬¬at the Account class Account.java and write a main method in a different class to briefly experiment with some instances of the Account class.
a.	Using the Account class as a base class, write two derived classes called SavingsAccount and CurrentAccount. A SavingsAccount object, in addition to the attributes of an Account object, should have an interest variable and a method which adds interest to the account. A CurrentAccount object, in addition to the attributes of an Account object, should have an overdraft limit variable. 
b.	Now create a Bank class which contains 1current and 1 saving account. 
c.	Now add interest of 10% to the above saving account created.
d.	Now increase the overdraft limit of the above account by 28%.
e.	Print the details of  the 2 accounts describing the account holder, account number , balance of the account.
f.	Note that the balance of an account may only be modified through the deposit(double) and withdraw(double) methods.